Remove himd_new that allocates for itself.
As the track metadata is no longer contained directly inside the himd
structure, it is feasible to allocate it on the stack, so remove the
allocating-and-initializing function himd_new. This allows to clean
up the mess introduced by having himd_close being the inverse operation
of himd_new instead of himd_open.