Determine current dataset number using atdata instead trkidx
Using trkidx is dangerous because there might be two of them. And
furthermore, taking the max number was wrong. It happens that
trkidx0f.hma is the backup/placeholder and trkidx00.hma is the current
one.

HiMDs seem to store the current "generation number" outside of the fs,
together with a master checksum. This generation number would be the
right value to use.