Work around broken Linux vfat file system
The vfat file system in Linux tries to emulate the case-preserving
case insensitivity of Windows. This includes presenting names with no
case information at all (because the case information is hidden in
reserved fields of the short name in special cases or otherwise given
by the long name) as all-uppercase. However, if the file system is
mounted using the UTF-8 charset (which nearly everyone does now),
the case insensitivity in the Linux kernel is suddenly disabled (they
don't support full unicode case insensitivity), so the files can
only be accessed using uppercase filenames.

Hi-MD media *should* be mounted as msdos or with an 8-bit iocharset
instead of UTF-8 to avoid the problem. Hi-MD does not need lowercase
or long file names at all, and also no national characters.

This patch works around the problem by always using uppercase
filenames.