Add filename case detection
As it is even more complicated then indicated in the last commit (vfat
driver can be configured [using shortname=lower] to present files
without case information as lower case, which *used* to be default, but
is no longer, now shortname=winnt seems to be default which uses a reserved
bit as case indicator. On plain msdos file systems, this bit is not set
which is interpreted as upper case), the code now detects whether upper case
file names work, and retries with lower case file names if upper case
file names fail.