Hack around qmake limitation - fixes translations
qmake's INSTALL mechanism can only install the default target, or files
already present in the source directory, as it searches these files
via VPATH and skips them if not found. This makes it impossible to
install autogenerated files like the qm files.

The current hack runs lrelease on qmake invocation to have qmake find
the files and generate install code.