NetMD downloading code
downloadhack.py is written in severe WFM quality, but should explain how
to use the download API. The EKB root key is missing, so it does not work
on your system.

Currently, it only works once per USB session with the MZ-RH1.
Creating unprotected tracks does not work with the IM-MZ899H, but that
might be a hardware limitation.