Simplify seek logic
The current heuristic for the multi-step waiting for seek completion and
play begin breaks on every new device the dump script is run on, and needed
extension everytime.

With the Sharp IM-MT899, *no* position change at all happens from the
"gotoTrack" call, so the first loop would never have ended, no matter how
soft the condition is expressed.

This patch should fix the problem once and for all, but it is not the nicest
way of coding it, as it will wait forever if we miss the second second of
audio (because of thrashing of the host machine, e.g.)