Force remaining size value to be unsigned 32-bit (Fixes #23)
Offset is size_t and sizeof() is size_t, which is fine on 32-bit
platforms, but not on 64-bit platforms; force the parameter of
qToLittleEndian() to be uint32_t, so we get sane results.

This only affected big-endian 64-bit archs (i.e. didn't affect
amd64), since qToLittleEndian() is a no-op on little endian archs.