Adds support for the MZ-M10 in both NetMD and Hi-MD modes
The MZ-M10 is the slightly more rare "Macintosh" variant of the MZ-RH910. It can be distinguished by its slightly different case finish (black body, silver border, black navigation wheel and track transport buttons). It appears identically to the MZ-RH910 over USB.

This patch adds the MZ-M10 to the identifiers for future convenience, and so that QHiMDTransfer additionally shows the M10 designation in the device dropdown.

For your reference, I've pulled the USB identifiers from my Mac's System Information.

In Hi-MD mode:

    Hi-MD:

      Product ID: 0x021c
      Vendor ID:  0x054c  (Sony Corporation)
      Version:  1.00
      Serial Number:  0200011069AB
      Speed:  Up to 12 Mb/sec
      Manufacturer: Sony
      Location ID:  0x14500000 / 12
      Current Available (mA): 500
      Current Required (mA):  500
      Extra Operating Current (mA): 0

And in NetMD mode:

    Net MD/Hi-MD:

      Product ID: 0x021b
      Vendor ID:  0x054c  (Sony Corporation)
      Version:  1.00
      Serial Number:  0200011069AB
      Speed:  Up to 12 Mb/sec
      Manufacturer: Sony
      Location ID:  0x14500000 / 15
      Current Available (mA): 500
      Current Required (mA):  500
      Extra Operating Current (mA): 0