Make NETMD_TITLE_RESPONSE_HEADER_SIZE a constant
This is the length of the title text header, in bytes. It's shared between libnetmd.c and trackinformation.c, and ois be used in libnetmd.c to handle request_disc_title returning the response's length, rather than just the text's length, which may be a bug, but is out of scope for this pull request.