Ensure pkg-config is correctly queried for `libgcrypt` and `libid3tag`
Previously, pkg-config was not invoked for `libgcrypt`, and both it and `libid3tag` was not required by all its dependent modules, and therefore non-standard installation locations for these dependencies would fail to be found. This updates the project files to correctly wire through the dependencies.

This patch does the following:

- `build/libgcrypt.pri`: Use qmake's `PKGCONFIG` variable rather than `LIBS`
- `libhimd/libhimd.pro`: Add `libid3tag.pri`, as it [is used by libhimd](https://github.com/linux-minidisc/linux-minidisc/blob/43a01190f819e44c5709017b889be9bf8766b989/libhimd/mp3tools.c#L26)
- `netmdcli/netmdcli.pro`: Add `libgcrypt.pri` as it [is used by netmdcli](https://github.com/linux-minidisc/linux-minidisc/blob/43a01190f819e44c5709017b889be9bf8766b989/netmdcli/netmdcli.c#L22)

I've tested this successfully on a macOS machine, but I believe it's acting in a fairly standard manner, and shouldn't change any results on other platforms.