Fix track duration for tracks longer than 1 hour
The duration value includes a "hours" field which needs to be read
and added to the minutes returned.

For reference, Web Minidisc / netmd-js properly reads 4 values and
gets the value right (the USB command seems to be slightly different,
but in any case, Web Minidisc shows the proper duration in its UI).

https://github.com/cybercase/netmd-js/blob/edc2401ea1840cecaeea42339b243f37109f8ece/src/netmd-interface.ts#L728